requirements.txt
aniso8601==10.0.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
Flask-RESTful==0.3.10
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pytest==8.3.4
python-dotenv==1.0.1
pytz==2025.1
redis==5.2.1
requests==2.32.3
six==1.17.0
urllib3==2.3.0
Werkzeug==3.1.3
